moved story_data to data folder
